requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-jazzmin==3.0.1
djangorestframework-simplejwt==5.5.1
PyMySQL==1.1.0
redis==5.0.1
celery==5.3.4
python-decouple==3.8
pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
drf-spectacular==0.27.0
reportlab==4.0.9
weasyprint==60.1
phonenumbers==8.13.0
django-audit-log==0.7.0
django-timezone-field==6.1.0
